changed Ui

+++ Application/src/main/kotlin/cga/exercise/game/Pages/MainMenuPage.kt
@@ -8,8 +8,8 @@ import cga.exercise.game.StaticResources
 class MainMenuPage(startButtonOnClick : ((Int,Int) -> Unit)) : LayoutBox(Relative(1f), Relative(1f), Center(), Center()) {
 
 
-    var updateFrequency = 60f
-    private val sliderUpdateFrequencyOnValueChanged = { f : Float -> updateFrequency = f * 599f + 1; updateFrequencyText.text = "$updateFrequency per sec" }
+    var updateFrequency = 60
+    private val sliderUpdateFrequencyOnValueChanged = { f : Float -> updateFrequency = (f * 599f + 1).toInt(); updateFrequencyText.text = "$updateFrequency per sec" }
 
     private val updateFrequencyText = Text("$updateFrequency per sec", 2.5f, StaticResources.standardFont,30f,TextMode.Right, false, PixelRight(60 + 213 + 10), Center(), StaticResources.fontColor1)
